Tried to make code work more like a genetic algorithm
Wanted to first remove the bad racecars, then look at the weights and biases of the good cars and store them for later.
Next we wanted to add the same amount of cars, that we removed, back into CarControllerList and give them all the same weights and biases, as the good cars that weren't deleted, as the new cars were created.
This all failed in the end, but it was due to an IndexOutOfBoundsException that we decided to stop trying to get it to work, since it clearly wasn't going to (i.e. it would have been a waste of time).